requirements.txt
mysql-connector-python
PyYAML
python-telegram-bot
torrent_parser
bencode.py